@@ -1,68 +1,186 @@
-﻿using CommunityToolkit.Mvvm.ComponentModel;
-using CommunityToolkit.Mvvm.Input;
-using MatchQuest.App.Views;
+﻿using CommunityToolkit.Mvvm.Input;
 using MatchQuest.Core.Interfaces.Services;
 using MatchQuest.Core.Models;
-using MatchQuest.Core.Data.Repositories;
-using Microsoft.Maui.Controls;
 using System.Collections.ObjectModel;
-using System.Threading.Tasks;
-using System.Linq;
+using CommunityToolkit.Mvvm.ComponentModel;
 
 namespace MatchQuest.App.ViewModels
 {
     public partial class HomeViewModel : BaseViewModel
     {
-        private readonly IAuthService _authService;
         private readonly GlobalViewModel _global;
         private readonly IMatchService _matchService;
-        private readonly ChatRepository _chatRepo;
+        private readonly IUserService _userService;
+        private readonly IChatService _chatService;
+        private readonly IReactionService _reactionService;
 
+        [ObservableProperty]
+        private bool isToastVisible;
+
+        [ObservableProperty]
+        private string toastMessage;
+
+        [ObservableProperty]
+        private Color toastBackgroundColor;
+        
+        // Token to handle resetting the timer if multiple toasts fire quickly
+        private CancellationTokenSource _toastCts;
+        
         private int CurrentUserId => _global?.Client?.Id ?? 0;
 
-        public ObservableCollection<User> Matches { get; } = new ObservableCollection<User>();
+        // Change the Matches collection type
+        public ObservableCollection<MatchChatItemViewModel> Matches { get; } = new();
 
-        public HomeViewModel(IAuthService authService, GlobalViewModel global, IMatchService matchService)
+        private List<MatchingScore> MatchPool { get; set; }
+        
+        [ObservableProperty]
+        private MatchingScore? currentMatch;
+
+        [ObservableProperty] private int currentMatchAge = 0;
+        
+        public bool HasMatch => CurrentMatch != null;
+        
+        public ImageSource CurrentProfileImage
+            => GetProfileImageSource(CurrentMatch?.Matcher);
+        
+        public ObservableCollection<string> GameTypeOptions { get; } =
+            new ObservableCollection<string>(
+                new[] { "All" }.Concat(Enum.GetNames(typeof(GameType)))
+            );
+        
+        [ObservableProperty]
+        private int selectedGameTypeIndex = 0;
+        
+        // Selected GameType (nullable = empty option)
+        public GameType? SelectedGameType =>
+            SelectedGameTypeIndex == 0
+                ? null
+                : (GameType)(SelectedGameTypeIndex - 1);
+
+        public HomeViewModel(
+            GlobalViewModel global, 
+            IMatchService matchService, 
+            IUserService userService, 
+            IChatService chatService, 
+            IReactionService reactionService)
         {
-            _authService = authService;
             _global = global;
             _matchService = matchService;
-            _chatRepo = new ChatRepository();
+            _chatService = chatService;
+            _userService = userService;
+            _reactionService = reactionService;
 
             LoadMatches();
+            LoadMatchPool();
         }
 
+        private void LoadMatchPool()
+        {
+            MatchPool = _userService.GetUserMatchPool(_global.Client, SelectedGameType);
+
+            if (MatchPool.Count > 0)
+            {
+                CurrentMatch = MatchPool.First();
+                CurrentMatchAge = CurrentMatch.Matcher.GetAge();
+                OnPropertyChanged(nameof(CurrentMatchAge));
+            }
+            else
+            {
+                CurrentMatch = null;
+            }
+
+            OnPropertyChanged(nameof(HasMatch));
+        }
+        
         private void LoadMatches()
         {
-            if (_global?.Client == null)
-                return;
+            if (_global?.Client == null) return;
 
-            var list = _matchService.GetAll(_global.Client.Id) ?? new System.Collections.Generic.List<User>();
+            var list = _matchService.GetAllMatchesFromUserId(_global.Client.Id);
             Matches.Clear();
 
             foreach (var u in list)
             {
-                // Populate UI-only preview property with the last message (if any)
-                u.LastMessagePreview = GetLastMessagePreview(u);
-                Matches.Add(u);
+                var preview = _chatService.GetLastMessagePreview(CurrentUserId, u);
+                var matchItem = new MatchChatItemViewModel(u, preview);
+                Matches.Add(matchItem);
             }
         }
-
-        private string? GetLastMessagePreview(User user)
+        
+        private void NextMatch()
         {
-            if (user == null || CurrentUserId == 0) return null;
+            if (CurrentMatch == null) return;
 
-            // Find the match row for the two users
-            var matchId = _chatRepo.GetMatchIdBetween(CurrentUserId, user.Id);
-            if (matchId == 0) return null;
+            var currentIndex = MatchPool?.IndexOf(CurrentMatch) ?? -1;
+            var lastIndex = (MatchPool?.Count ?? 0) - 1;
+            
+            if (currentIndex >= 0 && currentIndex < lastIndex)
+            {
+                CurrentMatch = MatchPool[currentIndex + 1];
+                OnPropertyChanged(nameof(CurrentMatchAge));
+            }
+            else
+            {
+                // We consumed the last match — refresh the pool to fetch new matches.
+                LoadMatchPool();
+            }
+        }
+        
+        private ImageSource GetProfileImageSource(User? user)
+        {
+            if (user == null || string.IsNullOrEmpty(user.ProfilePicture))
+                return ImageSource.FromFile("default_profile.png");
 
-            // Get or create chat for the match and fetch messages (repo returns ordered ASC)
-            var chatId = _chatRepo.GetOrCreateChatByMatchId(matchId);
-            var msgs = _chatRepo.GetMessagesByChatId(chatId);
-            if (msgs == null || msgs.Count == 0) return null;
+            try
+            {
+                var bytes = Convert.FromBase64String(user.ProfilePicture);
+                return ImageSource.FromStream(() => new MemoryStream(bytes));
+            }
+            catch
+            {
+                return ImageSource.FromFile("default_profile.png");
+            }
+        }
+        
+        private void ShowToast(string message, bool isSuccess)
+        {
+            // Cancel previous timer if exists
+            _toastCts?.Cancel();
+            _toastCts = new CancellationTokenSource();
 
-            // Return the last message text
-            return msgs.Last().MessageText;
+            // Set UI Properties
+            ToastMessage = message;
+            ToastBackgroundColor = isSuccess ? Colors.SeaGreen : Colors.IndianRed; // Nicer shades of Green/Red
+            IsToastVisible = true;
+
+            // Start timer on background thread
+            Task.Run(async () =>
+            {
+                try
+                {
+                    await Task.Delay(2000, _toastCts.Token);
+                    // If not cancelled, hide toast on Main Thread
+                    MainThread.BeginInvokeOnMainThread(() =>
+                    {
+                        IsToastVisible = false;
+                    });
+                }
+                catch (TaskCanceledException)
+                {
+                    // Task was cancelled because a new toast appeared; do nothing
+                }
+            });
+        }
+
+        private bool CreateReaction(bool isLike)
+        {
+            Reaction? reaction = null;
+            if (CurrentMatch != null)
+            {
+                reaction = _reactionService.CreateReaction(CurrentUserId, CurrentMatch.Matcher.Id, isLike);
+            }
+            
+            return reaction != null;
         }
 
         [RelayCommand]
@@ -81,18 +199,17 @@ namespace MatchQuest.App.ViewModels
 
             if (Shell.Current is not null)
             {
-                await Shell.Current.GoToAsync("Settings");
+                await Shell.Current.GoToAsync("UserProfile");
             }
         }
 
         // Called when a match is selected in the UI: set global selected match and navigate to Chat.
         [RelayCommand]
-        private async Task OpenChatFor(User? user)
+        private async Task OpenChatFor(MatchChatItemViewModel? matchItem)
         {
-            if (user is null)
-                return;
+            if (matchItem?.User is null) return;
 
-            _global.SelectedMatch = user;
+            _global.SelectedMatch = matchItem.User;
 
             // Ensure Shell is available (same pattern used elsewhere)
             if (Shell.Current is null)
@@ -108,6 +225,66 @@ namespace MatchQuest.App.ViewModels
             {
                 await Shell.Current.GoToAsync("Chat");
             }
+        }
+
+        [RelayCommand]
+        private void Like()
+        {
+            if (CurrentMatch == null)
+            {
+                return;
+            }
+
+            if (!CreateReaction(true))
+            {
+                ShowToast("Failed to like user.", false);
+            }
+            else
+            {
+                ShowToast($"Liked {CurrentMatch.Matcher.Name}!", true);
+            }
+            
+            LoadMatches();
+            NextMatch();
+        }
+
+        [RelayCommand]
+        private void Dislike()
+        {
+            if (CurrentMatch == null)
+            {
+                return;
+            }
+            
+            var success = CreateReaction(false);
+
+            if (!success)
+            {
+                ShowToast("Failed to dislike user.", false);
+            }
+            else
+            {
+                ShowToast($"Disliked {CurrentMatch.Matcher.Name}!", false);
+            }
+            
+            NextMatch();
+        }
+        
+        partial void OnCurrentMatchChanged(MatchingScore? value)
+        {
+            if (value != null)
+            {
+                currentMatchAge = value.Matcher.GetAge();
+            }
+            
+            OnPropertyChanged(nameof(CurrentMatchAge));
+            OnPropertyChanged(nameof(CurrentProfileImage));
+            OnPropertyChanged(nameof(HasMatch)); 
+        }
+        
+        partial void OnSelectedGameTypeIndexChanged(int value)
+        {
+            LoadMatchPool();
         }
     }
 }
