@@ -70,6 +70,9 @@ namespace MatchQuest.App.ViewModels
 
                 _global.Client = created;
 
+                // Persist logged-in user id so app can restore session
+                Preferences.Set("current_user_id", created.Id);
+
                 // Navigate to Home (or other next screen).
                 if (Shell.Current is null)
                 {
@@ -82,7 +85,7 @@ namespace MatchQuest.App.ViewModels
 
                 if (Shell.Current is not null)
                 {
-                    await Shell.Current.GoToAsync("Home");
+                    await Shell.Current.GoToAsync("RegisterProfilePicture");
                 }
             }
             catch (Exception ex)
