@@ -4,10 +4,9 @@ namespace MatchQuest.Core.Interfaces.Repositories
 {
     public interface IMatchRepository
     {
-        public User? Get(string email);
-        public User? Get(int id);
-        public List<User> GetAll();
-        public List<User> GetAll(int userId);
-        public User? Add(User client);
+        List<User> GetAllMatchesFromUserId(int userId);
+        Match? GetByUserIds(int userId, int userId2);
+
+        Match? CreateMatch(int userId1, int userId2);
     }
 }
