@@ -1,12 +1,20 @@
 namespace MatchQuest.Core.Models;
 
-public class Game
+public class Game : Entity
 {
-    public int Id { get; set; }
     public string Name { get; set; } = string.Empty;
     public GameType Type { get; set; }
-    public string Image { get; set; } = string.Empty;
     public bool Approved { get; set; }
-    public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
-    public DateTime UpdatedAt { get; set; }
+    public string Image { get; set; } = string.Empty;
+    
+    public Game(int id, string name, GameType type, bool approved, string image, DateTime createdAt, DateTime? updatedAt)
+    {
+        Id = id;
+        Name = name;
+        Type = type;
+        Approved = approved;
+        Image = image;
+        CreatedAt = createdAt;
+        UpdatedAt = updatedAt;
+    }
 }
