@@ -1,13 +1,16 @@
+using CommunityToolkit.Mvvm.ComponentModel;
+using CommunityToolkit.Mvvm.Input;
+using MatchQuest.Core.Helpers;
+using MatchQuest.Core.Interfaces.Repositories;
+using MatchQuest.Core.Models;
+
 namespace MatchQuest.Core.Models;
 
-public class Message
+public class Message : Entity
 {
-    public int Id { get; set; }
     public int ChatId { get; set; }
     public int SenderId { get; set; }
     public string MessageText { get; set; } = string.Empty;
-    public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
-    public DateTime UpdatedAt { get; set; }
 
     // UI-only flag - set by the viewmodel so XAML can show left/right correctly.
     // Not persisted to DB.
