@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics;
+using MatchQuest.Core.Helpers;
 using MatchQuest.Core.Interfaces.Repositories;
 using MatchQuest.Core.Interfaces.Services;
 using MatchQuest.Core.Models;
@@ -8,9 +9,13 @@ namespace MatchQuest.Core.Services
     public class UserService : IUserService
     {
         private readonly IUserRepository _userRepository;
-        public UserService(IUserRepository userRepository)
+        private readonly IGameService _gameService;
+        private readonly IReactionService _reactionService;
+        public UserService(IUserRepository userRepository, IGameService gameService, IReactionService reactionService)
         {
             _userRepository = userRepository;
+            _gameService = gameService;
+            _reactionService = reactionService;
         }
 
         public User? Get(string email)
@@ -45,18 +50,38 @@ namespace MatchQuest.Core.Services
             }
         }
         
-        public List<User> GetUsersWithMatchingGameType(int userId)
+        public List<MatchingScore> GetUserMatchPool(User matcher, GameType? gameType)
         {
             try
             {
-                var users = _userRepository.GetUsersWithMatchingGameType(userId);
-                return users;
+                var matchPool = new List<MatchingScore>();
+                
+                var gameTypeValue = gameType.HasValue ? (int)gameType.Value : -1;
+                
+                var users = _userRepository.GetUsersWithMatchingGameType(matcher.Id, gameTypeValue);
+                
+                foreach (var user in users)
+                {
+                    var mscore = AlgorithmHelper.CalculateMatchScore(user, matcher);
+                    matchPool.Add(mscore);
+
+                    user.Games = _gameService.ListByUserId(user.Id);
+                }
+                
+                // sort matchPool by Score descending
+                matchPool = matchPool.OrderByDescending(m => m.Score).ToList();
+                return matchPool;
             }
             catch (Exception ex)
             {
                 Debug.WriteLine($"UserService.GetUsersWithMatchingGameType: Exception: {ex}");
-                return new List<User>();
+                return [];
             }
+        }
+
+        public User? Update(User user)
+        {
+            return _userRepository.Update(user);
         }
     }
 }
